Break out merging of map/shuffle shards into separate step. Only run if not skipping the map phase.
